Optimise transforming calculators
